backend/pyproject.toml
[tool.poetry]
name = "devops-agent-backend"
version = "0.1.0"
description = "Backend for DevOps Agent Platform (FastAPI-based)"
authors = ["Your Team <devops@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.0"
pyyaml = "^6.0.2"
google-generativeai = "^0.3.2"
azure-mgmt-resource = "^25.0.0"
azure-mgmt-web = "^10.1.0"
azure-mgmt-containerservice = "^40.2.0"
azure-mgmt-compute = "^37.2.0"
azure-mgmt-network = "^30.2.0"
azure-identity = "^1.25.3"
pynacl = "^1.6.2"
sqlalchemy = {version = ">=2.0", extras = ["asyncio"]}
aiosqlite = "^0.20.0"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
